+++ app/reducers/grpc.js
@@ -8,26 +8,19 @@ import {
   GETSTAKEINFO_ATTEMPT, GETSTAKEINFO_FAILED, GETSTAKEINFO_SUCCESS,
   GETTICKETPRICE_ATTEMPT, GETTICKETPRICE_FAILED, GETTICKETPRICE_SUCCESS,
   GETACCOUNTS_ATTEMPT, GETACCOUNTS_FAILED, GETACCOUNTS_SUCCESS,
-  GETTRANSACTIONS_ATTEMPT, GETTRANSACTIONS_FAILED,  GETTRANSACTIONS_COMPLETE,
-  NEW_TRANSACTIONS_RECEIVED, CHANGE_TRANSACTIONS_FILTER,
   UPDATETIMESINCEBLOCK,
-  GETTICKETS_ATTEMPT, GETTICKETS_FAILED, GETTICKETS_COMPLETE, CLEAR_TICKETS,
-  GETTICKETS_PROGRESS, GETTICKETS_CANCEL,
   GETAGENDASERVICE_ATTEMPT, GETAGENDASERVICE_FAILED, GETAGENDASERVICE_SUCCESS,
-  RAWTICKETTRANSACTIONS_DECODED, CHANGE_TICKETS_FILTER,
+  RAWTICKETTRANSACTIONS_DECODED,
   GETMESSAGEVERIFICATIONSERVICE_ATTEMPT, GETMESSAGEVERIFICATIONSERVICE_FAILED, GETMESSAGEVERIFICATIONSERVICE_SUCCESS,
   GETVOTINGSERVICE_ATTEMPT, GETVOTINGSERVICE_FAILED, GETVOTINGSERVICE_SUCCESS,
   GETAGENDAS_ATTEMPT, GETAGENDAS_FAILED, GETAGENDAS_SUCCESS,
   GETVOTECHOICES_ATTEMPT, GETVOTECHOICES_FAILED, GETVOTECHOICES_SUCCESS,
   SETVOTECHOICES_ATTEMPT, SETVOTECHOICES_FAILED, SETVOTECHOICES_SUCCESS,
-  MATURINGHEIGHTS_CHANGED,
   GETBESTBLOCK_ATTEMPT, GETBESTBLOCK_FAILED, GETBESTBLOCK_SUCCESS,
   STARTWALLETSERVICE_ATTEMPT,
   STARTWALLETSERVICE_FAILED, STARTWALLETSERVICE_SUCCESS, GETTREASURY_BALANCE_SUCCESS, RESET_TREASURY_BALANCE,
-  FETCHMISSINGSTAKETXDATA_ATTEMPT, FETCHMISSINGSTAKETXDATA_SUCCESS, FETCHMISSINGSTAKETXDATA_FAILED,
-  GETSTARTUPTRANSACTIONS_SUCCESS,
-  GETALLAGENDAS_SUCCESS, GETALLAGENDAS_FAILED,
-  ABANDONTRANSACTION_ATTEMPT, ABANDONTRANSACTION_SUCCESS, ABANDONTRANSACTION_FAILED
+  ABANDONTRANSACTION_ATTEMPT, ABANDONTRANSACTION_SUCCESS, ABANDONTRANSACTION_FAILED,
+  GETALLAGENDAS_SUCCESS, GETALLAGENDAS_FAILED
 } from "../actions/ClientActions";
 import { DAEMONSYNCED, WALLETREADY } from "../actions/DaemonActions";
 import { NEWBLOCKCONNECTED } from "../actions/NotificationActions";
@@ -41,6 +34,14 @@ import {
   CLOSEWALLET_SUCCESS
 } from "actions/WalletLoaderActions";
 import { GETACCOUNTMIXERSERVICE_SUCCESS, RUNACCOUNTMIXER_SUCCESS, STOPMIXER_SUCCESS } from "actions/AccountMixerActions";
+import {
+  FETCHMISSINGSTAKETXDATA_ATTEMPT, FETCHMISSINGSTAKETXDATA_SUCCESS, FETCHMISSINGSTAKETXDATA_FAILED,
+  GETTRANSACTIONS_ATTEMPT, GETTRANSACTIONS_FAILED,  GETTRANSACTIONS_COMPLETE,
+  GETTICKETS_ATTEMPT, GETTICKETS_FAILED, GETTICKETS_COMPLETE, CLEAR_TICKETS,
+  GETTICKETS_PROGRESS, GETTICKETS_CANCEL, CHANGE_TICKETS_FILTER,
+  MATURINGHEIGHTS_CHANGED, GETSTARTUPTRANSACTIONS_SUCCESS,
+  NEW_TRANSACTIONS_RECEIVED, CHANGE_TRANSACTIONS_FILTER
+} from "actions/TransactionActions";
 
 export default function grpc(state = {}, action) {
   switch (action.type) {
@@ -376,8 +377,6 @@ export default function grpc(state = {}, action) {
   case GETTRANSACTIONS_COMPLETE:
     return {
       ...state,
-      minedTransactions: action.minedTransactions,
-      unminedTransactions: action.unminedTransactions,
       transactions: [ ...action.unminedTransactions, ...action.minedTransactions ],
       noMoreTransactions: action.noMoreTransactions,
       lastTransaction: action.lastTransaction,
@@ -412,7 +411,6 @@ export default function grpc(state = {}, action) {
       minedTransactions: [],
       unminedTransactions: [],
       transactions: [],
-      lastTransaction: null,
       noMoreTransactions: false
     };
   case FETCHMISSINGSTAKETXDATA_ATTEMPT:
@@ -676,8 +674,8 @@ export default function grpc(state = {}, action) {
         types: [],
         direction: null
       },
-      recentRegularTransactions: Array(),
-      recentStakeTransactions: Array(),
+      recentRegularTransactions: [],
+      recentStakeTransactions: [],
       ticketBuyerService: null,
       transactionsSinceLastOpened: null,
       votingService: null,
