+++ app/reducers/snackbar.js
@@ -29,7 +29,6 @@ import {
 } from "../actions/StakePoolActions";
 import {
   ABANDONTRANSACTION_SUCCESS, ABANDONTRANSACTION_FAILED,
-  NEW_TRANSACTIONS_RECEIVED,
   GETSTARTUPWALLETINFO_FAILED,
   SEEDCOPIEDTOCLIPBOARD
 } from "../actions/ClientActions";
@@ -45,6 +44,7 @@ import {
   GETWALLETSEEDSVC_FAILED,
   SPVSYNC_FAILED
 } from "actions/WalletLoaderActions";
+import { NEW_TRANSACTIONS_RECEIVED } from "actions/TransactionActions";
 import {
   TRZ_TOGGLEPINPROTECTION_SUCCESS, TRZ_TOGGLEPINPROTECTION_FAILED,
   TRZ_TOGGLEPASSPHRASEPROTECTION_SUCCESS, TRZ_TOGGLEPASSPHRASEPROTECTION_FAILED,
